requirements.txt
# Requirements for RAG System with Gemini
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pyyaml==6.0.1
python-dotenv==1.0.0
PySide6>=6.8.0
requests==2.31.0
aiohttp==3.9.1  # For async Gemini API calls

# PDF Support (optional but recommended)
PyPDF2==3.0.1

# Alternative PDF library (choose one)
# pdfplumber==0.10.3

# Optional for production
# openai==1.3.5
# anthropic==0.8.0
# faiss-cpu==1.7.4
# chromadb==0.4.18
# sentence-transformers==2.2.2
